+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - [#118] Make implicitly used init container images explicit and configurable
-   - this release adds a mandatory ConfigMap `k8s-ces-additional-images` which contains additionally used images
+   - this release adds a mandatory ConfigMap `k8s-dogu-operator-additional-images` which contains additionally used images
    - see the [operations docs](docs/operations/installing_operator_into_cluster_en.md) for more information
 
 ## [v0.35.1] - 2023-08-31